Implement documentation generation script
Documentation generation code was added to the go:generate calls. Its
implementation is a bit hacky as the tool needs a valid GOPATH, but
it seems to work for our use case.

Unfortunately it's only able to generate HTML but it can be integrated
into asciidoc with a simple include statemenet.

Relates to: #37