Implement git file templates
With this commit the GitRepo CRD was extended so it can hold a map of
file and content. Currently it's only tested to create files in the
root of a repository and not deeper in the file tree.

Also each cluster will now create an empty template file in its tenant
repository.

Relates to: #36