Adjust CRD doc templates
The html that gets generated with the default templates for the
`gen-crd-api-reference-docs` tool don't contain any classes in the
table tags. HTML classes are necessary to get the right look and feel
for our Antora sites. Because the HTML is simply pased through so we
have to handle the styling ourselves.