Fix reconcile loop with GitRepo updates
Updating the git repository every on every reconcile triggers a loop.
That's because the gitRepo controller adds fields that aren't set in the
gitRepoTemplates of the tenants and clusters. So it will always apply
changes.

The operator will permanently reconcile all tenants and clusters as they
have the Gitrepository as their secondary objects. That triggers and
endless loop of reconciles.

The spec of the GitRepo object must not be changed in the reconcile
function! GitRepo objects will be created from templates contained
within the cluster or tenant objects.