Add deletion logic
This commit add logic to delete Git and Vault resources. There are two
kinds of deletion for each resource:

* soft deletion -> the resource can easily be restored if necessary
* hard deletion -> the resource is completely deleted

Also there's an annotation that is set by default, that prevents the
operator from deleting resources. Please see the docs for more info.

Resources will be deleted as soon as the k8s resource belonging to it is
deleted!