Fix gitlab.com detection due to CloudFlare
The check to determine if it's a gitlab instance or not will query the
`/users/sign_in` path. If that path is not found it's not a GitLab
instance. Gitlab.com is using some CloudFlare browser detection on that
path, which works very well and blocks that request from the operator.

Now the operator will determine that it's a GitLab instance if the
endpoint contains `gitlab.com`.