Avoid reconcile loop
The previous implementation caused the CRs to be updated with each
reconcile. Although no actual changes were applied the ResourceVersion
increased every time. This lead to the status not being set correctly.
And of course excessive reconciling.

By uptdating the status first and catching conflicts, we can avoid that.
This bug was only triggered for newly created GitRepos.