Add details about where errors occurred
During troubleshooting issues with Lieutenant, we noticed that some
error messages did not contain enough information to find out which step
of a pipeline failed.

This commit tries to improve the errors reported back to the user by
making sure all errors returned in a `Result` contain some information
about the precise step that triggered it.

Signed-off-by: Manuel Hutter <manuel@hutter.io>