Fix issue that blocked status updates
Update on the main resource does not update the status, but it
does fetch the current version of that status from k8s. We
currently use the orginal object in this update call, which has
its status overwritten.  We need to copy the resource when
updating to make sure that we can update both the resource and
 its status.