Reconcile Tenants and Clusters when their ServiceAccounts change
We add ServiceAccounts as an owned resource when setting up the tenant
and cluster controllers. This ensures that a reconcile of the owning
tenant or cluster is triggered if its associated ServiceAccount is
deleted.

Additionally, we change the owner reference on the tenant ServiceAccount
to be a controller reference so that the reconcile for a tenant is
actually triggered when its ServiceAccount is deleted. Notably,
reconciles aren't triggered for owned resources, if the owner reference
isn't a controller reference (cf. documentation for
[controllerutil.SetControllerReference] and
[controllerutil.SetOwnerReference])

[controllerutil.SetControllerReference]: https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/controller/controllerutil#SetControllerReference
[controllerutil.SetOwnerReference]: https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/controller/controllerutil#SetOwnerReference