Implement ServiceAccount token secret generation for clusters
We add logic to create a ServiceAccount token secret for the
ServiceAccount created for each Cluster when the feature is enabled on
the operator.

The secret is created as documented in
https://kubernetes.io/docs/reference/access-authn-authz/service-accounts-admin/#to-create-additional-api-tokens

We set a controller reference on the generated secret, so a cluster
reconcile is triggered if the secret is deleted.