Implement ServiceAccount token secret generation for tenants
We add logic to create a ServiceAccount token secret for the
ServiceAccount created for each Tenant when the feature is enabled on
the operator.

The secret is created as documented in
https://kubernetes.io/docs/reference/access-authn-authz/service-accounts-admin/#to-create-additional-api-tokens

We set a controller reference on the generated secret, so a tenant
reconcile is triggered if the secret is deleted.