Remove unnecessary `url.PathEscape` for `GetNamespace`
The new version of the GitLab library URL-escapes the provided repo path
in `GetNamespace`, so we must not escape it ourselves anymore, as we
otherwise get a double-URL-escaped value for the request, which fails.

Also add a test case for creating a project in a subgroup. We also
verify the request URL's `RawPath` field, to ensure that the subgroup id
lookup query is done with the correctly escaped group path.