Make git repo creation behaviour configurable
This commit introduces a new field `CreationPolicy` for all managed
resoruces. This field configures how Lieutenant handles the creation of
Git repositories.

By default Lieutenant will not manage git repositories it did not create
itself. This behaviour can be changed by setting the Creation Policy to
`adopt`, which will allows Lieutenant to manage existing repositories.
This can be helpful for migrations or recovery.